change in block

+++ quorum-examples/vagrant/bootstrap.sh
@@ -7,28 +7,52 @@ apt-get update
 sudo snap install solc
 apt-get install -y build-essential unzip libdb-dev libleveldb-dev libsodium-dev zlib1g-dev libtinfo-dev sysvbanner wrk software-properties-common default-jdk maven
 
+# install constellation
+CVER="0.3.2"
+CREL="constellation-$CVER-ubuntu1604"
+wget -q https://github.com/jpmorganchase/constellation/releases/download/v$CVER/$CREL.tar.xz
+tar xfJ $CREL.tar.xz
+cp $CREL/constellation-node /usr/local/bin && chmod 0755 /usr/local/bin/constellation-node
+rm -rf $CREL
+
 # install golang
 GOREL=go1.9.3.linux-amd64.tar.gz
-wget -q https://dl.google.com/go/${GOREL}
-tar xfz ${GOREL}
-mv go /usr/local/go
-rm -f ${GOREL}
+wget -q https://dl.google.com/go/$GOREL
+tar xfz $GOREL
+if [ ! -e "/usr/local/go" ]; then
+    mv go /usr/local/go
+fi
+rm -f $GOREL
 PATH=$PATH:/usr/local/go/bin
 echo 'PATH=$PATH:/usr/local/go/bin' >> /home/vagrant/.bashrc
 
 # make/install quorum
+#rm -r probablistic-blockchain
+if [ -e "probablistic-blockchain" ]; then
+	rm -r probablistic-blockchain
+#	rm -r /usr/local/bin/geth
+	#rm -r /usr/local/bin/bootnode
+	rm -r /home/vagrant/quorum-examples
+fi
 git clone https://tarasalman:video1234@bitbucket.org/tarasalman/probablistic-blockchain.git/
-#pushd probablistic-blockchain >/dev/null
-sudo chmod 777 -R probablistic-blockchain
-cd probablistic-blockchain
+pushd probablistic-blockchain >/dev/null
+	#git checkout tags/v2.0.1
 make all
 cp build/bin/geth /usr/local/bin
 cp build/bin/bootnode /usr/local/bin
 popd >/dev/null
 
+# install Porosity
+wget -q https://github.com/jpmorganchase/quorum/releases/download/v1.2.0/porosity
+mv porosity /usr/local/bin && chmod 0755 /usr/local/bin/porosity
 
-sudo chgrp -R vagrant probablistic-blockchain
-sudo chown -R vagrant probablistic-blockchain
+# copy examples
+cp -r /vagrant/examples /home/vagrant/quorum-examples
+chown -R vagrant:vagrant /home/vagrant/probablistic-blockchain /home/vagrant/quorum-examples
 
-echo 'Quorum source ready'
+# done!
+banner "Quorum"
+echo
+echo 'The Quorum vagrant instance has been provisioned. Examples are available in ~/quorum-examples inside the instance.'
+echo "Use 'vagrant ssh' to open a terminal, 'vagrant suspend' to stop the instance, and 'vagrant destroy' to remove it."
